// aas-web-ui/src/composables/Client/CompanyLookup/types/company.ts
import type { types as aasTypes } from '@aas-core-works/aas-core3.1-typescript'

// Todo: types should be autogenerated with every new deployment of the the company lookup service and be provide as a package

export interface AdministrativeInformation {
  version?: string | null
  revision?: string | null
  creator?: aasTypes.Reference | null
}

export interface ProtocolInformation {
  href: string
  // Todo: BE is restrictive, we need allowed enums so that FE and BE can use the same validation
  endpointProtocol?: string | null
}

export interface Endpoint {
  protocolInformation: ProtocolInformation
  // Todo: BE is restrictive, we need allowed enums so that FE and BE can use the same validation
  interface: string
}

export interface CompanyDescriptor {
  idShort?: string | null
  name?: string | null
  domain: string
  nameOptions?: string[] | null
  assetIdRegexPatterns?: string[] | null
  description?: aasTypes.LangStringTextType[] | null
  displayName?: aasTypes.LangStringTextType[] | null
  administration?: AdministrativeInformation | null
  endpoints: Endpoint[]
}